refactor: move to tasty process commander
